# chatgpt-actions.openapi.yaml
openapi: 3.1.0
info:
  title: Anna Personal Memory
  description: >-
    Complete Anna REST API for accounts, payment channels, transactions,
    completed life-log events, reminders, collections, and constrained MongoDB
    operations. Dates are local calendar dates in YYYY-MM-DD format.
  version: 1.1.0
servers:
  - url: https://anna-production-bfb9.up.railway.app
security:
  - bearerAuth: []
paths:
...
  /v1/collections:
    get:
      operationId: listCollections
      summary: List accessible collections
      responses:
        "200":
          description: Accessible managed and flexible collections.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CollectionListResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
    post:
      operationId: createCollection
      summary: Create a flexible collection
      description: Managed collection names cannot be created through this operation.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCollectionRequest"
      responses:
        "200":
          description: The collection already existed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateCollectionResponse"
        "201":
          description: The collection was created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateCollectionResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
  /v1/accounts:
    post:
      operationId: createAccount
      summary: Create a financial account
      description: >-
        Create an account only after the user explicitly confirms its identity,
        kind, and currency. Transactions can reference the returned account ID.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAccountRequest"
      responses:
        "201":
          description: The account was created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateAccountResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
  /v1/payment-channels:
    post:
      operationId: createPaymentChannel
      summary: Create a payment channel
      description: >-
        Create an intermediary such as TrueMoney Wallet or LINE Pay. A payment
        channel is distinct from the account charged for the transaction.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePaymentChannelRequest"
      responses:
        "201":
          description: The payment channel was created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreatePaymentChannelResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
...
  /v1/reminders:
    post:
      operationId: createReminder
      summary: Create a recurring reminder
      description: >-
        Create one recurring reminder rule. Ask the user to resolve ambiguous
        weekdays, start date, timezone, or preparation lead time before calling.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateReminderRequest"
      responses:
        "200":
          description: The reminder already existed and was returned.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateReminderResponse"
        "201":
          description: The reminder was created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateReminderResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
  /v1/reminders/completions:
    post:
      operationId: completeReminder
      summary: Complete one reminder step
      description: >-
        Complete one phase of one dated reminder occurrence. occurrenceOn is
        the scheduled occurrence date, including when completing preparation
        before that date. Ask if the occurrence is ambiguous.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompleteReminderRequest"
      responses:
        "200":
          description: This completion already existed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WriteResponse"
        "201":
          description: The reminder step was completed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WriteResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
...
  /v1/events/search:
    post:
      operationId: searchEvents
      summary: Search completed life-log events
      description: >-
        Search with at least one structured filter. Resolve relative dates such
        as last Friday into absolute local dates before calling. Text search is
        normalized token matching, not fuzzy or semantic search.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EventSearchRequest"
      responses:
        "200":
          description: Matching events.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventSearchResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
  /v1/events/{id}:
    get:
      operationId: getEvent
      summary: Get one event
      parameters:
        - name: id
          in: path
          required: true
          description: Event MongoDB object identifier.
          schema:
            type: string
            pattern: "^[a-fA-F0-9]{24}$"
      responses:
        "200":
          description: The event.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
    patch:
      operationId: updateEvent
      summary: Correct an event
      description: >-
        Apply merge-patch semantics to mutable event fields. Null clears an
        optional field. Server-owned fields and original source text cannot be
        changed.
      parameters:
        - name: id
          in: path
          required: true
          description: Event MongoDB object identifier.
          schema:
            type: string
            pattern: "^[a-fA-F0-9]{24}$"
      requestBody:
        required: true
        content:
          application/merge-patch+json:
            schema:
              $ref: "#/components/schemas/UpdateEventRequest"
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateEventRequest"
      responses:
        "200":
          description: The updated event.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
    delete:
      operationId: deleteEvent
      summary: Delete an event
      description: Delete only the event; linked transactions are unchanged.
      parameters:
        - name: id
          in: path
          required: true
          description: Event MongoDB object identifier.
          schema:
            type: string
            pattern: "^[a-fA-F0-9]{24}$"
      responses:
        "204":
          description: The event was deleted.
        default:
          $ref: "#/components/responses/ErrorResponse"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  responses:
    ErrorResponse:
      description: The request failed.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    CollectionInfo:
      type: object
      additionalProperties: false
      required: [name, managed]
      properties:
        name:
          type: string
        managed:
          type: boolean
    CollectionListResponse:
      type: object
      required: [data, meta]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/CollectionInfo"
        meta:
          type: object
          required: [count]
          properties:
            count:
              type: integer
    CreateCollectionRequest:
      type: object
      additionalProperties: false
      required: [name]
      properties:
        name:
          type: string
          pattern: "^[a-z][a-z0-9_]{0,62}$"
    CreateCollectionResponse:
      type: object
      additionalProperties: false
      required: [data, created]
      properties:
        data:
          $ref: "#/components/schemas/CollectionInfo"
        created:
          type: boolean
    CreateAccountRequest:
      type: object
      additionalProperties: false
      required: [id, name, kind]
      properties:
        id:
          type: string
          pattern: "^[a-z0-9]+(?:-[a-z0-9]+)*$"
          maxLength: 100
          description: Stable lowercase kebab-case identifier.
        name:
          type: string
          minLength: 1
          maxLength: 200
        kind:
          type: string
          enum: [bank_account, credit_card, cash, wallet, other]
        currency:
          type: string
          pattern: "^[A-Z]{3}$"
          default: THB
    Account:
      type: object
      additionalProperties: false
      required: [id, name, kind, currency, active]
      properties:
        id:
          type: string
        name:
          type: string
        kind:
          type: string
          enum: [bank_account, credit_card, cash, wallet, other]
        currency:
          type: string
        active:
          type: boolean
    CreateAccountResponse:
      type: object
      additionalProperties: false
      required: [data, created]
      properties:
        data:
          $ref: "#/components/schemas/Account"
        created:
          type: boolean
    CreatePaymentChannelRequest:
      type: object
      additionalProperties: false
      required: [id, name]
      properties:
        id:
          type: string
          pattern: "^[a-z0-9]+(?:-[a-z0-9]+)*$"
          maxLength: 100
          description: Stable lowercase kebab-case identifier.
        name:
          type: string
          minLength: 1
          maxLength: 200
    PaymentChannel:
      type: object
      additionalProperties: false
      required: [id, name, active]
      properties:
        id:
          type: string
        name:
          type: string
        active:
          type: boolean
    CreatePaymentChannelResponse:
      type: object
      additionalProperties: false
      required: [data, created]
      properties:
        data:
          $ref: "#/components/schemas/PaymentChannel"
        created:
          type: boolean
...
    CreateReminderRequest:
      type: object
      additionalProperties: false
      required: [id, title, weekdays]
      properties:
        id:
          type: string
          pattern: "^[a-z0-9]+(?:-[a-z0-9]+)*$"
          maxLength: 100
          description: Stable lowercase kebab-case identifier.
        title:
          type: string
          minLength: 1
          maxLength: 300
        timezone:
          type: string
          default: Asia/Bangkok
          description: IANA timezone.
        weekdays:
          type: array
          minItems: 1
          maxItems: 7
          uniqueItems: true
          items:
            type: string
            enum:
              - monday
              - tuesday
              - wednesday
              - thursday
              - friday
              - saturday
              - sunday
        startsOn:
          type: string
          format: date
          description: Omit to start today in the reminder timezone.
        preparation:
          $ref: "#/components/schemas/ReminderPreparation"
    ReminderPreparation:
      type: object
      additionalProperties: false
      required: [title, leadDays]
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 300
        leadDays:
          type: integer
          minimum: 1
          maximum: 30
          description: >-
            Start showing preparation this many days before the occurrence and
            repeat it daily through the day before while incomplete.
    CreateReminderResponse:
      type: object
      required: [created, data]
      properties:
        created:
          type: boolean
        data:
          type: object
          required: [id, title, timezone, weekdays, startsOn, active]
          properties:
            id:
              type: string
            title:
              type: string
            timezone:
              type: string
            weekdays:
              type: array
              items:
                type: string
            startsOn:
              type: string
              format: date
            preparation:
              $ref: "#/components/schemas/ReminderPreparation"
            active:
              type: boolean
    CompleteReminderRequest:
      type: object
      additionalProperties: false
      required: [reminderId, occurrenceOn, phase]
      properties:
        reminderId:
          type: string
          description: Stable reminder rule identifier.
        occurrenceOn:
          type: string
          format: date
          description: Scheduled date of the reminder occurrence.
        phase:
          type: string
          enum: [preparation, occurrence]
...
    EventType:
      type: string
      enum:
        - company-event
        - social-event
        - meal-gathering
        - appointment-completed
        - travel
        - celebration
        - personal-milestone
        - work-milestone
        - health-fitness
        - pet
        - family
        - general-memory
    EventFinancialContext:
      type: object
      additionalProperties: false
      required: [currency]
      minProperties: 2
      properties:
        currency:
          type: string
          pattern: "^[A-Z]{3}$"
        totalValueMinor:
          type: integer
          minimum: 0
        allowanceMinor:
          type: integer
          minimum: 0
        coveredByOthersMinor:
          type: integer
          minimum: 0
        personalPaymentMinor:
          type: integer
          minimum: 0
          description: Zero only when the source explicitly says the user paid nothing.
    EventAttributeScalar:
      oneOf:
        - type: string
          maxLength: 1000
        - type: number
        - type: boolean
        - type: "null"
    EventAttributeValue:
      oneOf:
        - $ref: "#/components/schemas/EventAttributeScalar"
        - type: array
          maxItems: 20
          items:
            $ref: "#/components/schemas/EventAttributeScalar"
    EventAttributes:
      type: object
      maxProperties: 32
      description: >-
        Guarded event-specific extension data. Keys cannot start with a dollar
        sign or contain a dot. Values are short JSON scalars or scalar arrays.
        The listed keys are examples; descriptive additional keys are allowed.
      properties:
        restaurantChain:
          $ref: "#/components/schemas/EventAttributeValue"
        occasion:
          $ref: "#/components/schemas/EventAttributeValue"
        department:
          $ref: "#/components/schemas/EventAttributeValue"
      additionalProperties:
        $ref: "#/components/schemas/EventAttributeValue"
...
    EventSearchRequest:
      type: object
      additionalProperties: false
      minProperties: 1
      properties:
        occurredFrom:
          type: string
          format: date
        occurredTo:
          type: string
          format: date
        createdFrom:
          type: string
          format: date-time
        createdTo:
          type: string
          format: date-time
        eventTypes:
          type: array
          uniqueItems: true
          items:
            $ref: "#/components/schemas/EventType"
        location:
          type: string
          maxLength: 500
        people:
          type: array
          maxItems: 50
          items:
            type: string
            maxLength: 200
        tags:
          type: array
          maxItems: 20
          items:
            type: string
            maxLength: 50
        text:
          type: string
        sort:
          type: string
          enum: [occurred_desc, occurred_asc, created_desc]
          default: occurred_desc
        limit:
          type: integer
          minimum: 1
          maximum: 100
          default: 50
    UpdateEventRequest:
      type: object
      additionalProperties: false
      minProperties: 1
      properties:
        occurredOn:
          type: string
          format: date
        occurredAt:
          oneOf:
            - type: string
              format: date-time
            - type: "null"
        timezone:
          type: string
        title:
          type: string
          minLength: 1
          maxLength: 300
        eventType:
          $ref: "#/components/schemas/EventType"
        location:
          oneOf:
            - type: string
              minLength: 1
              maxLength: 500
            - type: "null"
        people:
          oneOf:
            - type: array
              maxItems: 50
              items:
                type: string
            - type: "null"
        description:
          oneOf:
            - type: string
              maxLength: 5000
            - type: "null"
        tags:
          oneOf:
            - type: array
              maxItems: 20
              items:
                type: string
            - type: "null"
        financialContext:
          oneOf:
            - $ref: "#/components/schemas/EventFinancialContext"
            - type: "null"
        relatedTransactionIds:
          oneOf:
            - type: array
              maxItems: 20
              uniqueItems: true
              items:
                type: string
                pattern: "^[a-fA-F0-9]{24}$"
            - type: "null"
        attributes:
          oneOf:
            - $ref: "#/components/schemas/EventAttributes"
            - type: "null"
    EventDocument:
      type: object
      description: Managed event document encoded as MongoDB Extended JSON.
      properties:
        _id:
          description: MongoDB Extended JSON object identifier.
        schemaVersion:
          type: integer
        occurredOn:
          type: string
          format: date
        occurredAt:
          description: Optional MongoDB Extended JSON date-time.
        timezone:
          type: string
        title:
          type: string
        eventType:
          $ref: "#/components/schemas/EventType"
        location:
          description: Location string or null.
        people:
          type: array
          items:
            type: string
        description:
          description: Event description string or null.
        tags:
          type: array
          items:
            type: string
        financialContext:
          description: Financial context object or null.
        relatedTransactionIds:
          type: array
          items:
            description: MongoDB Extended JSON object identifier.
        attributes:
          $ref: "#/components/schemas/EventAttributes"
        source:
          type: object
          properties:
            type:
              type: string
            requestId:
              type: string
              format: uuid
            requestHash:
              type: string
            rawText:
              type: string
        createdAt:
          description: MongoDB Extended JSON date-time.
        updatedAt:
          description: MongoDB Extended JSON date-time.
      additionalProperties: true
    EventResponse:
      type: object
      required: [data]
      properties:
        data:
          $ref: "#/components/schemas/EventDocument"
    EventSearchResponse:
      type: object
      required: [data, meta]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/EventDocument"
        meta:
          type: object
          required: [count]
          properties:
            count:
              type: integer
    WriteResponse:
      type: object
      required: [data]
      properties:
        data:
          type: object
          required: [id, created]
          properties:
            id:
              description: MongoDB Extended JSON identifier or string identifier.
            created:
              type: boolean
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
            message:
              type: string